port: Mojo 1.0.0 stable — five compile classes + the GPU teardown-order hang
Baseline 236/236 on mojo-1.0.0b2.dev2026051806; now 236/236 across all 9
suites on Mojo 1.0.0 (2026-08-11) + MAX 26.5.0. Every change is the compiler
refusing something that was unsound, plus one runtime ordering bug it exposed.

1. Recursive structs vs conditional Deinitable (template.Value, template.Node,
   json.JsonValue). List[T] is Deinitable only where T is, and these hold a
   List of themselves; the synthesised destructor cannot close the cycle.
   Declare Deinitable + an explicit __deinit__ (fields still destroyed
   implicitly). Probe-verified.
2. Aliasing on self-assign through a slice (app.mojo, serve.mojo): the leading-'/'
   strip read the target immutably while constructing it. Temporary + transfer.
3. Implicit Int -> UInt dropped: perf_counter_ns() is Int, the logger takes
   UInt ns. Convert at the boundary (logger, middleware chain, 3 examples).
4. dlopen callables are borrow-checked; struct fields may not expose AnyOrigin;
   imm -> mut pointer widening rejected (gpu.mojo, gpu_store.mojo) — taken
   from the private line's dev2026080106 port, which fixed the same classes.
5. DeviceContext/DeviceBuffer moved to max.gpu.host.

The hang (test_queue_api, open since the August port): Mojo destroys struct
fields in declaration order. GpuQueue declared its DeviceContext before the
DeviceBuffer it owned, so the context died first; the next DeviceContext()+
allocation in the same process then deadlocked in the driver. Reordered the
fields and pinned the order in an explicit __deinit__. Six probes localise it
(second context alone OK; second context after ctx-before-buffer teardown
hangs; buffer-before-ctx OK).

+++ examples/chat/main.mojo
@@ -65,7 +65,7 @@ struct ChatApp(DispatchHandler, Movable):
         return self.templates.render(String("chat.html"), ctx)
 
     def __call__(mut self, req: Request) raises -> Response:
-        var t0 = perf_counter_ns()
+        var t0 = UInt(perf_counter_ns())  # Mojo 1.0: no implicit Int -> UInt; the logger takes UInt ns
         var resp: Response
 
         if req.method == "GET" and req.path == "/":
